Split failover GraphQL API into separate module
