+++ cartridge/webui.lua
@@ -12,6 +12,7 @@ local api_config = require('cartridge.webui.api-config')
 local api_issues = require('cartridge.webui.api-issues')
 local api_vshard = require('cartridge.webui.api-vshard')
 local api_topology = require('cartridge.webui.api-topology')
+local api_failover = require('cartridge.webui.api-failover')
 local api_ddl = require('cartridge.webui.api-ddl')
 local gql_types = require('cartridge.graphql.types')
 
@@ -48,6 +49,7 @@ local function init(httpd)
 
     -- Basic topology operations
     api_topology.init(graphql)
+    api_failover.init(graphql)
 
     -- Replication warnings and other problems
     api_issues.init(graphql)
